Completed Dev on Localized Recreation Map and ajusted Facilities API to work with the map.
